refactoring modit/dit and initializaitons
