+++ src/exojax/spec/dit.py
@@ -59,10 +59,11 @@ def npgetix(x,xv):
     return cont,index
 
 @jit
-def inc3D_givenx(w,cx,ix,y,z,xv,yv,zv):
+def inc3D_givenx(a,w,cx,ix,y,z,xv,yv,zv):
     """The lineshape distribution matrix = integrated neighbouring contribution for 3D (memory reduced sum) but using given contribution and index for x .
     
     Args:
+        a: lineshape density array (jnp.array)
         w: weight (N)
         cx: given contribution for x 
         ix: given index for x 
@@ -86,7 +87,6 @@ def inc3D_givenx(w,cx,ix,y,z,xv,yv,zv):
     cy,iy=getix(y,yv)
     cz,iz=getix(z,zv)
 
-    a=jnp.zeros((len(xv),len(yv),len(zv)))
     a=index_add(a,joi[ix,iy,iz],w*(1-cx)*(1-cy)*(1-cz))
     a=index_add(a,joi[ix,iy+1,iz],w*(1-cx)*cy*(1-cz))
     a=index_add(a,joi[ix+1,iy,iz],w*cx*(1-cy)*(1-cz))
@@ -99,155 +99,7 @@ def inc3D_givenx(w,cx,ix,y,z,xv,yv,zv):
     return a
 
 @jit
-def inc3D(w,x,y,z,xv,yv,zv):
-    """The lineshape distribution matrix = integrated neighbouring contribution for 3D (memory reduced sum).
-    
-    Args:
-        w: weight (N)
-        x: x values (N)
-        y: y values (N)
-        z: z values (N)
-        xv: x grid
-        yv: y grid
-        zv: z grid            
-        
-    Returns:
-        lineshape distribution matrix (integrated neighbouring contribution for 3D)
-        
-    Note:
-        This function computes \sum_n w_n fx_n \otimes fy_n \otimes fz_n, 
-        where w_n is the weight, fx_n, fy_n, and fz_n are the n-th NCFs for 1D. 
-        A direct sum uses huge RAM. 
-        In this function, we use jax.lax.scan to compute the sum
-        
-    Example:
-        >>> N=10000
-        >>> xv=jnp.linspace(0,1,11) #grid
-        >>> yv=jnp.linspace(0,1,13) #grid
-        >>> zv=jnp.linspace(0,1,17) #grid
-        >>> w=np.logspace(1.0,3.0,N)
-        >>> x=np.random.rand(N)
-        >>> y=np.random.rand(N)
-        >>> z=np.random.rand(N)
-        >>> val=inc3D(w,x,y,z,xv,yv,zv)
-        >>> #the comparision with the direct sum
-        >>> valdirect=jnp.sum(nc3D(x,y,z,xv,yv,zv)*w,axis=3)
-        >>> #maximum deviation
-        >>> print(jnp.max(jnp.abs((val-valdirect)/jnp.mean(valdirect)))*100,"%") #%
-        >>> 5.520862e-05 %
-        >>> #mean deviation
-        >>> print(jnp.sqrt(jnp.mean((val-valdirect)**2))/jnp.mean(valdirect)*100,"%") #%
-        >>> 8.418057e-06 %
-    """
-
-    cx,ix=getix(x,xv)
-    cy,iy=getix(y,yv)
-    cz,iz=getix(z,zv)
-
-    a=jnp.zeros((len(xv),len(yv),len(zv)))
-    a=index_add(a,joi[ix,iy,iz],w*(1-cx)*(1-cy)*(1-cz))
-    a=index_add(a,joi[ix,iy+1,iz],w*(1-cx)*cy*(1-cz))
-    a=index_add(a,joi[ix+1,iy,iz],w*cx*(1-cy)*(1-cz))
-    a=index_add(a,joi[ix+1,iy+1,iz],w*cx*cy*(1-cz))
-    a=index_add(a,joi[ix,iy,iz+1],w*(1-cx)*(1-cy)*cz)
-    a=index_add(a,joi[ix,iy+1,iz+1],w*(1-cx)*cy*cz)
-    a=index_add(a,joi[ix+1,iy,iz+1],w*cx*(1-cy)*cz)
-    a=index_add(a,joi[ix+1,iy+1,iz+1],w*cx*cy*cz)
-
-    return a
-
-
-@jit
-def xsvector(nu_lines,sigmaD,gammaL,S,nu_grid,sigmaD_grid,gammaL_grid,dLarray):
-    """Cross section vector (DIT/3D version, default)
-    
-    The original code is rundit in [addit package](https://github.com/HajimeKawahara/addit)
-
-    Args:
-       nu_lines: line center (Nlines)
-       sigmaD: Gaussian STD (Nlines)
-       gammaL: Lorentzian half width (Nlines)
-       S: line strength (Nlines)
-       nu_grid: linear wavenumber grid
-       sigmaD_grid: sigmaD grid
-       gammaL_grid: gammaL grid
-       dLarray: ifold/dnu (ifold=1,..,Nfold) array
-
-    Returns:
-       Cross section in the linear nu grid
-
-    Note:
-       This version uses inc3D. So, nu grid is also auto-differentiable. However, be careful for the precision of wavenumber grid, because inc3D uses float32 in JAX/GPU. For instance, consider to use dfnus=nus-np.median(nus) and dfnu_lines=mdbCO.nu_lines-np.median(nus) instead of nus (nu_grid) and nu_lines, to mitigate the problem. 
-
-    Example:
-       >>> dfnus=nus-np.median(nus)
-       >>> dfnu_lines=nu_lines-np.median(nus)
-       >>> dnus=nus[1]-nus[0]
-       >>> Nfold=3
-       >>> dLarray=jnp.linspace(1,Nfold,Nfold)/dnus 
-       >>> xs=xsvector3D(nu_lines,sigmaD,gammaL,Sij,nus,sigmaD_grid,gammaL_grid,dLarray)
-
-    """
-    Ng_nu=len(nu_grid)
-    Ng_sigmaD=len(sigmaD_grid)
-    Ng_gammaL=len(gammaL_grid)
-    
-    log_sigmaD=jnp.log(sigmaD)
-    log_gammaL=jnp.log(gammaL)
-    
-    log_sigmaD_grid = jnp.log(sigmaD_grid)
-    log_gammaL_grid = jnp.log(gammaL_grid)
-    dnu = (nu_grid[-1]-nu_grid[0])/(Ng_nu-1)
-    k = jnp.fft.rfftfreq(2*Ng_nu,dnu)
-    val=inc3D(S,nu_lines,log_sigmaD,log_gammaL,nu_grid,log_sigmaD_grid,log_gammaL_grid)
-
-    valbuf=jnp.vstack([val,jnp.zeros_like(val)])
-    fftval = jnp.fft.rfft(valbuf,axis=0)
-    #vk=voigt_kernel(k, sigmaD_grid,gammaL_grid)
-    #vk=f1_voigt_kernel(k, sigmaD_grid,gammaL_grid, dnu)
-    vk=folded_voigt_kernel(k, sigmaD_grid,gammaL_grid, dLarray)
-
-    fftvalsum = jnp.sum(fftval*vk,axis=(1,2))
-    xs=jnp.fft.irfft(fftvalsum)[:Ng_nu]/dnu
-    return xs
-
-
-@jit
-def xsmatrix(nu_lines,sigmaDM,gammaLM,SijM,nu_grid,dgm_sigmaD,dgm_gammaL,dLarray):
-    """Cross section matrix for xsvector (DIT/reduced memory version)
-
-    Args:
-       nu_lines: line center (Nlines)
-       sigmaDM: doppler sigma matrix in R^(Nlayer x Nline)
-       gammaLM: gamma factor matrix in R^(Nlayer x Nline)
-       SijM: line strength matrix in R^(Nlayer x Nline)
-       nu_grid: linear wavenumber grid
-       dgm_sigmaD: DIT Grid Matrix for sigmaD R^(Nlayer, NDITgrid)
-       dgm_gammaL: DIT Grid Matrix for gammaL R^(Nlayer, NDITgrid)
-       dLarray: ifold/dnu (ifold=1,..,Nfold) array
-      
-    Return:
-       cross section matrix in R^(Nlayer x Nwav)
-
-    """
-    NDITgrid=jnp.shape(dgm_sigmaD)[1]
-    Nline=len(nu_lines)
-    Mat=jnp.hstack([sigmaDM,gammaLM,SijM,dgm_sigmaD,dgm_gammaL])
-    def fxs(x,arr):
-        carry=0.0
-        sigmaD=arr[0:Nline]
-        gammaL=arr[Nline:2*Nline]
-        Sij=arr[2*Nline:3*Nline]
-        sigmaD_grid=arr[3*Nline:3*Nline+NDITgrid]
-        gammaL_grid=arr[3*Nline+NDITgrid:3*Nline+2*NDITgrid]
-        arr=xsvector(nu_lines,sigmaD,gammaL,Sij,nu_grid,sigmaD_grid,gammaL_grid,dLarray)
-        return carry, arr
-    
-    val,xsm=scan(fxs,0.0,Mat)
-    return xsm
-    
-@jit
-def xsvector_np(cnu,indexnu,sigmaD,gammaL,S,nu_grid,sigmaD_grid,gammaL_grid, dLarray):
+def xsvector(cnu,indexnu,sigmaD,gammaL,S,nu_grid,sigmaD_grid,gammaL_grid, dLarray):
     """Cross section vector (DIT/2D+ version; default)
     
     The original code is rundit in [addit package](https://github.com/HajimeKawahara/addit)
@@ -287,7 +139,9 @@ def xsvector_np(cnu,indexnu,sigmaD,gammaL,S,nu_grid,sigmaD_grid,gammaL_grid, dLa
     log_gammaL_grid = jnp.log(gammaL_grid)
     dnu = (nu_grid[-1]-nu_grid[0])/(Ng_nu-1)
     k = jnp.fft.rfftfreq(2*Ng_nu,dnu)
-    val=inc3D_givenx(S,cnu,indexnu,log_sigmaD,log_gammaL,nu_grid,log_sigmaD_grid,log_gammaL_grid)
+    
+    lsda=jnp.zeros((len(nu_grid),len(log_sigmaD_grid),len(log_gammaL_grid)))
+    val=inc3D_givenx(lsda,S,cnu,indexnu,log_sigmaD,log_gammaL,nu_grid,log_sigmaD_grid,log_gammaL_grid)
 
     valbuf=jnp.vstack([val,jnp.zeros_like(val)])
     fftval = jnp.fft.rfft(valbuf,axis=0)
@@ -297,7 +151,7 @@ def xsvector_np(cnu,indexnu,sigmaD,gammaL,S,nu_grid,sigmaD_grid,gammaL_grid, dLa
     return xs
 
 @jit
-def xsmatrix_np(cnu, indexnu, sigmaDM,gammaLM,SijM,nu_grid,dgm_sigmaD,dgm_gammaL,dLarray):
+def xsmatrix(cnu, indexnu, sigmaDM,gammaLM,SijM,nu_grid,dgm_sigmaD,dgm_gammaL,dLarray):
     """Cross section matrix (DIT/2D+ version)
 
     Args:
@@ -325,7 +179,7 @@ def xsmatrix_np(cnu, indexnu, sigmaDM,gammaLM,SijM,nu_grid,dgm_sigmaD,dgm_gammaL
         Sij=arr[2*Nline:3*Nline]
         sigmaD_grid=arr[3*Nline:3*Nline+NDITgrid]
         gammaL_grid=arr[3*Nline+NDITgrid:3*Nline+2*NDITgrid]
-        arr=xsvector_np(cnu,indexnu,sigmaD,gammaL,Sij,nu_grid,sigmaD_grid,gammaL_grid,dLarray)
+        arr=xsvector(cnu,indexnu,sigmaD,gammaL,Sij,nu_grid,sigmaD_grid,gammaL_grid,dLarray)
         return carry, arr
     
     val,xsm=scan(fxs,0.0,Mat)
